feat: front end updates
