+++ front-svelte/src/AppConfig.js
@@ -1,7 +1,8 @@
 let AppConfig = {
   SITE_FOOTER_AUTHOR: 'philbjern',
   SITE_FOOTER_AUTHOR_URL: 'https://github.com/philbjern',
-  API_URL: 'http://localhost:8088/api'
+  API_URL: 'http://localhost:8088/api',
+  NOTIFICATION_TIMEOUT: 1500,
 }
