+++ front-svelte/src/NotificationType.js
@@ -1,0 +1,7 @@
+const NotificationType = {
+  INFO: 'MEDIUM',
+  SUCCESS: 'SUCCESS',
+  ERROR: 'ERROR'
+}
+
+export default NotificationType
