--- front-svelte/src/routes/project/+page.svelte
@@ -1,5 +1,0 @@
-<script>
-  let projectId = 1
-</script>
-
-<h1>Project </h1>
