--- front-svelte/src/routes/project/[slug]/+page.svelte
@@ -1,5 +1,0 @@
-<script>
-  export let slug
-</script>
-
-<h1>Project {slug}</h1>
