+++ front-svelte/src/routes/styles.css
@@ -1355,6 +1355,12 @@ input[type="text"]:focus {
   color: var(--color-white);
 }
 
+	
+.dark-mode .nav-item.active {
+  color: white;
+  border-bottom: 2px solid white;
+}
+
 /* MEDIA QUERIES */
 
 @media screen and (min-width: 1450px) {
